Address comments in review, 4
