--- src/stringUtils.h
@@ -1,28 +1,0 @@
-/*
- * Copyright 2020 Andrei Pangin
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-#ifndef _STRINGUTILS_H
-#define _STRINGUTILS_H
-
-#include <string.h>
-
-
-class StringUtils {
-  public:
-    static bool endsWith(const std::string& s, const char* suffix, size_t suffixlen);
-    static void replace(std::string& s, char c, const char* replacement, size_t rlen);
-};
-
-#endif // _STRINGUTILS_H
